requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-smorest
python-dotenv